services/wordcloud/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
fastprogress==1.0.0
feedparser==6.0.2
grpcio==1.35.0
grpcio-tools==1.35.0
idna==2.10
itsdangerous==1.1.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
mecab-python3==1.0.3
neologdn==0.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.11.13
requests==2.25.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2
urllib3==1.26.3
Werkzeug==1.0.1
wordcloud==1.8.1
